requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.3
bidict==0.23.1
blinker==1.9.0
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudevents==1.11.0
cloudpathlib==0.21.1
cmudict==1.0.32
colorama==0.4.6
confection==0.1.5
cymem==2.0.11
dataclasses-json==0.6.7
dateparser==1.2.1
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
filelock==3.18.0
filetype==1.2.0
Flask==3.1.0
Flask-Cors==4.0.0
Flask-Limiter==3.12
Flask-SocketIO==5.5.1
frozenlist==1.6.0
fsspec==2025.5.1
functions-framework==3.8.2
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-secret-manager==2.23.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
greenlet==3.2.2
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.1
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.59
langchain-google-genai==2.0.10
langchain-openai==0.3.17
langchain-text-splitters==0.3.8
langcodes==3.5.0
langsmith==0.3.42
language_data==1.3.0
libcst==1.8.0
limits==5.2.0
lxml==5.4.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
murmurhash==1.0.13
mypy_extensions==1.1.0
narwhals==1.35.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
oauthlib==3.2.2
openai==1.72.0
ordered-set==4.1.0
orjson==3.10.18
packaging==24.2
pandas==2.2.3
pillow==11.2.1
plotly==5.18.0
pluggy==1.6.0
preshed==3.0.10
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pypdf==5.5.0
pyperclip==1.9.0
pyphen==0.17.2
pytest==8.4.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.2
PyYAML
PyYAML-ft==7.0.1
redis==6.1.0
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.24.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
setuptools==80.9.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
streamlit==1.45.1
streamlit-elements==0.1.0
sympy==1.14.0
tenacity==9.1.2
textstat==0.7.7
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.52.3
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.3.0
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.0
zipp==3.22.0
zstandard==0.23.0
